CHROMIUM: cpuquiet: Don't hold cpuquiet_lock in cpuquiet_wake_quiesce_cpu()
Calls to cpuquiet_wake_quiesce_cpu() are already serialized by the
governors and attempts to change the governor should be waiting for
completion of any outstanding requests in the governor's ->stop()
callback (e.g. cancel_work_sync() in runnables_stop()), so there is no
need to hold the cpuquiet_lock here.

BUG=chrome-os-partner:45826
TEST=Build and boot on Smaug, no more lockdep splat.

Change-Id: Iadd6af4da630825af15dfe6c3f9e06e8480c3764
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/302226
Reviewed-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Harsh Shandilya <msfjarvis@gmail.com>